py/oogabooga/price_checker_bot/requirements.txt
python-telegram-bot
requests
python-dotenv
web3
matplotlib
nest-asyncio